refactor: split CommandDispatcher into per-concern handler modules
