+++ lib/browserctl/server/handlers/page_lifecycle.rb
@@ -1,0 +1,29 @@
+# frozen_string_literal: true
+
+module Browserctl
+  class CommandDispatcher
+    module Handlers
+      module PageLifecycle
+        private
+
+        def cmd_open_page(req)
+          session = @global_mutex.synchronize do
+            @pages[req[:name]] ||= PageSession.new(@browser.create_page)
+          end
+          session.page.go_to(req[:url]) if req[:url]
+          { ok: true, name: req[:name] }
+        end
+
+        def cmd_close_page(req)
+          session = @global_mutex.synchronize { @pages.delete(req[:name]) }
+          session&.page&.close
+          { ok: true }
+        end
+
+        def cmd_list_pages(_req)
+          { pages: @global_mutex.synchronize { @pages.keys } }
+        end
+      end
+    end
+  end
+end
